refactor: remove component subfolders (#184)
(and unused components)
