+++ src/stories/components/Prompt.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { Button } from '@views/components/common/Button/Button';
-import type { PromptDialogProps } from '@views/components/common/Prompt/Prompt';
-import PromptDialog from '@views/components/common/Prompt/Prompt';
+import { Button } from '@views/components/common/Button';
+import type { PromptDialogProps } from '@views/components/common/Prompt';
+import PromptDialog from '@views/components/common/Prompt';
 import Text from '@views/components/common/Text/Text';
 import React from 'react';
