+++ src/stories/components/ScheduleListItem.stories.tsx
@@ -1,6 +1,6 @@
 import { UserSchedule } from '@shared/types/UserSchedule';
 import type { Meta, StoryObj } from '@storybook/react';
-import ScheduleListItem from '@views/components/common/ScheduleListItem/ScheduleListItem';
+import ScheduleListItem from '@views/components/common/ScheduleListItem';
 import useSchedules from '@views/hooks/useSchedules';
 import React from 'react';
