+++ src/stories/components/calendar/CalendarBottomBar.stories.tsx
@@ -2,7 +2,7 @@ import { Course, Status } from '@shared/types/Course';
 import Instructor from '@shared/types/Instructor';
 import { getCourseColors } from '@shared/util/colors';
 import type { Meta, StoryObj } from '@storybook/react';
-import CalendarBottomBar from '@views/components/calendar/CalendarBottomBar/CalendarBottomBar';
+import CalendarBottomBar from '@views/components/calendar/CalendarBottomBar';
 import React from 'react';
 
 const exampleGovCourse: Course = new Course({
