--- src/views/components/calendar/CalendarCourseBlock/CalendarCourseMeeting.module.scss
@@ -1,34 +1,0 @@
-.component {
-    display: flex;
-    padding: 7px 7px 9px 7px;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 5px;
-    flex: 1 0 0;
-    align-self: stretch;
-    border-radius: 4px;
-    background: #cbd5e1;
-    .content {
-        display: flex;
-        gap: 7px;
-        .course-detail {
-            display: flex;
-            flex-direction: column;
-            gap: 5px;
-            width: 154px;
-            .course {
-                font-size: 16px;
-                line-height: 16px;
-                font-weight: 600;
-                white-space: nowrap;
-                overflow: hidden;
-                text-overflow: ellipsis;
-            }
-            .time-and-location {
-                font-size: 11px;
-                line-height: 11px;
-                font-weight: 400;
-            }
-        }
-    }
-}
