--- src/views/components/common/Card/Card.module.scss
@@ -1,10 +1,0 @@
-@use 'src/views/styles/colors.module.scss';
-
-.card {
-    background: colors.$white;
-    border: 0.5px dotted #c3cee0;
-    box-sizing: border-box;
-
-    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-    border-radius: 8px;
-}
