+++ src/views/components/common/ConflictsWithWarning.tsx
