+++ src/views/components/common/CourseStatus.tsx
