+++ src/views/components/common/PopupCourseBlock.tsx
