+++ src/views/components/common/ScheduleListItem.tsx
