+++ src/views/components/common/ScheduleTotalHoursAndCourses.tsx
